+++ HISTORY.md
@@ -2,6 +2,7 @@
 ## v0.0.2
 - support default command to search: ```./fofa port=80```
 - search support -o param to write to file: ```./fofa search -o a.txt port=80```
+- add global verbose option to debug: ```./fofa --verbose search port=80```
 
 ## v0.0.1 initial release
 - add search/account subcommand